requirements.txt
tensorflow
keras
streamlit
numpy
Pillow
opencv-python